+++ .gitignore
@@ -13,4 +13,3 @@ out/
 .settings
 .classpath
 .factorypath
-native-image-result/
