+++ native-image-items/config/predefined-classes-config.json
@@ -1,0 +1,8 @@
+[
+  {
+    "type":"agent-extracted",
+    "classes":[
+    ]
+  }
+]
+
