+++ native-image-items/config/proxy-config.json
@@ -1,0 +1,2 @@
+[
+]
