+++ native-image-items/config/resource-config.json
@@ -1,0 +1,5 @@
+{
+  "resources":{
+  "includes":[]},
+  "bundles":[]
+}
