+++ native-image-items/config/serialization-config.json
@@ -1,0 +1,6 @@
+{
+  "types":[
+  ],
+  "lambdaCapturingTypes":[
+  ]
+}
