+++ src/main/kotlin/com/hrothwell/anime/Main.kt
@@ -1,14 +1,13 @@
 package com.hrothwell.anime
 
+import com.github.ajalt.clikt.core.subcommands
 import com.hrothwell.anime.commands.Anime
 import com.hrothwell.anime.commands.Login
+import com.hrothwell.anime.commands.Random
 
 object Main {
   @JvmStatic
   fun main(args: Array<String>) {
-//    Anime().main(args)
-    Login().main(args)
-  // TODO adding subcommands seemed to break native image as it no longer created reflection-config.json. Tried several things with no luck.
-    // see if there is some other way to set subcommands, or just use flags to dictate logic?
+    Anime().subcommands(Login(), Random()).main(args)
   }
 }
