+++ src/main/kotlin/com/hrothwell/anime/commands/Anime.kt
@@ -12,58 +12,12 @@ import com.hrothwell.anime.util.AnimeUtil.Companion.RED
 import kotlinx.serialization.decodeFromString
 
 class Anime : CliktCommand(
-  help = "pick a random anime from your MAL lists",
+  help = "CLI for interacting with MAL",
   epilog = """
     extended user guide, source code, etc - https://github.com/hrothwell/anime-cli
   """.trimIndent()
 ) {
+  // entry point
+  override fun run() = Unit
 
-  private val possibleListStatusValues = ListStatus.values().map { it.malValue }.toTypedArray()
-  private val user by option("-u", "--user-name", help = """
-    user name. if not provided this will default to "user_name" value in "${AnimeUtil.home}/anime-cli/mal-secret.json"
-  """.trimIndent())
-
-  private val list by option(
-    "-l",
-    "--list-status",
-    help = "list to select from"
-  ).choice(choices = possibleListStatusValues)
-    .default("plan_to_watch")
-
-  override fun run() {
-    getRandomAnime(user, list)
-  }
-
-  fun getRandomAnime(user: String?, list: String) {
-
-    val clientSecrets = AnimeUtil.getUserSecrets()
-
-    val headers = "X-MAL-CLIENT-ID" to clientSecrets.client_id
-    val listStatus = "status" to list
-    val limit = "limit" to 1000
-    val userPathParam = user ?: clientSecrets.user_name
-
-    val request =
-      Fuel.get(path = "https://api.myanimelist.net/v2/users/$userPathParam/animelist", parameters = listOf(listStatus, limit))
-        .appendHeader(headers)
-
-    // TODO check for errors before getting result here
-    //   (also why does Fuel do this this way it feels weird to call .third)
-    val response = request.response()
-    AnimeUtil.handlePotentialHttpErrors(response.second)
-
-    try {
-      val json = String(response.third.get())
-      val result = AnimeUtil.jsonReader.decodeFromString<MALUserListResponse>(json)
-      echo("Anime selected: ${result.data.random().node.title}")
-    } catch (t: Throwable) {
-      echoError("couldn't read the response from MAL: $t")
-      echoError("Stack trace of error: \n")
-      t.printStackTrace(System.err)
-    }
-  }
-
-  private fun echoError(msg: String) {
-    echo("$RED $msg", err = true)
-  }
 }
